requirements/base.txt
attrs==24.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2>=3.1.4
MarkupSafe==3.0.2
outcome==1.3.0.post0
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio-websocket==0.11.1
trio==0.27.0
Werkzeug==3.0.4
wsproto==1.2.0